app/__init__.py
# app/__init__.py

from flask import Flask
from config.settings import Config
from flask_sqlalchemy import SQLAlchemy

app = Flask(__name__)
app.config.from_object(Config)
db = SQLAlchemy(app)

from app import routes, models